html/browsers/browsing-the-web/navigating-across-documents/010-timeout.html

<!doctype html>
<title>Link with onclick form submit to javascript: url with delayed (timeout) document.write and href navigation</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe id="test" name="test"></iframe>
<form target="test" action="javascript:setTimeout(() => document.write('<script>parent.postMessage(&quot;write&quot;, &quot;*&quot;)</script>'), 2000);"></form>
<a target="test" onclick="document.forms[0].submit()" href="href.html">Test</a>
<script>
/*---
specid: navigating-across-documents
steps: [6, 11, 12]
info: |
  Cancel any preexisting but not yet mature attempt to navigate browsingContext, including canceling any instances of the fetch algorithm started by those attempts. If one of those attempts has already created and initialized a new Document object, abort that Document also. (Navigation attempts that have matured already have session history entries, and are therefore handled during the update the session history with the new page algorithm, later.)

  Step 12 contains https://html.spec.whatwg.org/multipage/browsing-the-web.html#javascript-protocol

  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#concept-form-submit
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#submit-get-action
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#plan-to-navigate
  - https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#planned-navigation

description: |
  This test expects the form's `action`, which contains a sync XMLHttpRequest (delayed by 2 seconds) to be aborted by navigating to the anchor's `href`.
---*/
var t = async_test(undefined, {
  timeout: 4000
});
var iframe = document.getElementsByTagName("iframe")[0];
var form = document.getElementsByTagName("form")[0];

t.step(function() {
  document.getElementsByTagName("a")[0].click();
});

onmessage = t.step_func(function(e) {
  assert_equals(e.data, "href");
  t.done();
  // iframe.parentNode.removeChild(iframe);

});
</script>
